requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.26.0
monai>=1.3.0
gunicorn==21.2.0
python-dotenv==1.0.0 